16.14 migration

+++ canflood/model/modcom.py
@@ -1816,7 +1816,7 @@ class Model(ComWrkr,
                            sorted(ar)[1]/10, #dummy value for zero (take the second smallest value and divide by 10)
                            ar) 
         
-        df['ari'] = ar_ari.astype(np.int32)
+        df['ari'] = ar_ari.astype(int)
         
         return 
     
@@ -2634,7 +2634,7 @@ class DFunc(ComWrkr, #damage function or DFunc handler
  
         #typeset it
         try:
-            ddf2 = ddf2.astype(np.float)
+            ddf2 = ddf2.astype(float)
         except Exception as e:
             raise Error('failed to typsset the ddf for \'%s\' w/ \n    %s'%(self.tabn, e))
         
@@ -2796,8 +2796,8 @@ class DFunc(ComWrkr, #damage function or DFunc handler
         
         #check it
         try:
-            ddf.columns = ddf.columns.astype(np.float)
-            ddf = ddf.astype(np.float)
+            ddf.columns = ddf.columns.astype(float)
+            ddf = ddf.astype(float)
         except Exception as e:
             """
             wont work if you have a very messy library
